Added methods for retrieving and posting data to/from database

+++ src/main/java/se/val18/miniprojekt/repo/Hit.java
@@ -3,28 +3,28 @@ package se.val18.miniprojekt.repo;
 import java.time.LocalDateTime;
 
 public class Hit {
-    long id;
-    long wordId;
-    long domainId;
+    int id;
+    int wordId;
+    int domainId;
     String context;
     LocalDateTime createTime;
 
-    public Hit(long wordId, long domainId, String context) {
+    public Hit(int wordId, int domainId, String context) {
         this.id = id;
         this.wordId = wordId;
         this.domainId = domainId;
         this.context = context;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public long getWordId() {
+    public int getWordId() {
         return wordId;
     }
 
-    public long getDomainId() {
+    public int getDomainId() {
         return domainId;
     }
